[test] Ensure faces have meaningful foreground/background for noninteractive tests
When running tests noninteractive (for example when in CI, or by calling "make
test") Emacs doesn't have faces initialized.  This causes tests to fail, as
every colour substitution is to "unspecified-fg" or "unspecified-bg".
Apparently, this is a featrue, see
https://lists.gnu.org/archive/html/help-gnu-emacs/2024-02/msg00095.html.

This change will ensure the faces used in tests have some distinct values when
running noninteractively.

It also turns out that with a current implementation Emacs-28 highlight faces
are not applied for diffs.  Quick investigation shows that
`difftastic--ansi-color-add-background`
called twice: first time with `ansi-color-underline`, second time with
`(:foregorund "something")`.  Unfortunately, the former is used to detect if
the given face should be updated by `difftastic` to be a highlight
face.  Marking tests as failing for time being, when there's a demand will
revisit if it's worth the effort.