refactor: Allow to bind to symbols in difftastic--with-temp-files
This makes the macro closer to what reqular `let*` would do, that is to bind
the file-bufs when they are declared. Switch to `let*` has been dictated by a
need to delete temporary files that could have been created while binding and
one of the following valueforms signals an error.