feat: Add extensions to temporary files when comparing buffers
This is to help `difftastic` to automatically select language when comparing
buffers that don't have underlaying files.

Use output of `difft --list-languages` to find all extensions that `difftastic`
knows about then map them against `auto-mode-alist` to find relevant major
modes and extend that list with `-ts-modes`. Only return the first extension,
that matches a mode as it will be appended to a temporary file that will be
read by `difftastic`.

closes: #5