feat: Custom completion for language overrides for difftastic arguments
This implements a custom completion function, that helps user to enter valid
GLOB:LANGUAGE patterns when specifying `--override` difftastic argument in
`difftastic--with-extra-arguments` transient.

The idea is that anything in user typed string is tried to be splitted with
semicolon. Everything before semicolon is a GLOB (with a default of
*). Everything after a colon is a (a part of) LANGUAGE. The completions
proposed match current GLOB:part-of-LANGUAGE.