refactor!: Don't save buffer if it is modified
When a buffer to be compared visits a file, don't save it, but create a
temporary file with `buffer-string` instead.

This is in preparation for `difftasitc-buffer-file` where the whole point is to
compare an modified buffer content with a visiting file.

BREAKING CHANGE: When a buffer visits a file, and the buffer is modified don't
  save the buffer before running difftastic. User needs to manually save the
  buffer to a file if the temporary file names in the output are too
  obfuscated.