fix: When a modified buffer visits a file favor file extensions
When a temporary file is created for a buffer difftastic tries to guess an
extension for the temporary file. Before change in
a6c9b5b37691e632f4d1a7f59c388757fe6f3335, there were only two cases. Either a
buffer has been visiting a file, in such a case buffer was saved to a file
before feeding the file to difftastic, or a buffer was not visiting a file in
which case a temporary file was created and extension for it was guessed based
on the buffer mode. However, the change introduced a third case: a modified
buffer that visits a file. I think it's better to use the visited file
extension (if any) when creating a temporary file in such a case.