feat!: Prevent buffers command to select the same buffer
This is to ensure there's no confusion when `difftastic-rerun` is called after
`difftastic-file-buffer`. The latter command uses the same buffer, albeit for
one of arguments it uses file visited by the buffer while for the other it
creates a new temporary file with the buffer contents. When `difftastic-rerun`
is called it makes assumption when the buffer is the same that it should use
that version of arguments.

Should the same buffer be picked up for `difftastic-buffer` that would create a
bad user experience when calling `difftastic-rerun as it will show diff between
different contents (as compared to a first call).

BREAKING CHANGE: Buffers for command `difftastic-buffers` and
  `difftastic--buffer` now have to be different.