feat: Add command difftastic-forge-pullreq-show-diff
The new command `difftastic-forge-pullreq-show-diff` is meant to show a diff
for a pull request like. That is a diff for `base...head` range. Both `base`
and `head` are guessed based on a pull request either at point or associated
with a current `forge` buffer. When no pull request can be guessed it asks for
a branches like a `forge` would when creating a pull request.

The companion function `difftastic-forge-create-pulreq-show-diff` to be
used in `forge-edit-post-hook`.

N.B., this functionality heavily relies on `forge`, hence it needs to be
installed. If it is not available, the regular `difftastic-git-diff-range` can
be used to manually select `base...head` range.