refactor: Use process-lines instead of shell-command-to-string
This follows a recommendation in `shell-command-to-string`. Indeed, shell is
not required when calling `difft`. More over, it has a benefit that in a case
of error when calling difftastic (not installed or something else) such an
error will be propagated to user (albeit a bit cryptic) instead of silently
swallowed.