test: Don't use org-mode in tests
Thing that triggered this change is the following failure of byte compilation
on Emacs-28:

test/difftastic.t.el:14:1: Error: Invalid function: org-assert-version

`org-mode` has been only used to have a buffer in a mode that does not have a
suggestion.  However this is not the best example as there may appear a
tree-sitter grammar for `.org` files and difftastic may add support for
it. Switch to use difftastic-mode instead it doesn't seem reasonable to write
a tree-sitter grammar for difftastic output. I hope I am not terribly far off
with the latter claim ¯\_(ツ)_/¯.

Also add missing (require 'treesit) to accommodate for lexical binding of
`treesit-auto-install-grammar` in tests.