feat!: Use arguments saved in transient for a difft call
This makes the values saved in `transient` sticky for all `difft`
invocations. This feature may be useful, for example to present more context
lines by default, by specifying `--context=5` and then saving the state of the
transient with `C-x C-s`.

BREAKING CHANGE: Any arguments saved in `transient` are now used for all
  `difft` invocations (unless explicitly overridden by calling a difftastic
  command with double prefix argument). Change
  `difftastic-use-transient-arguments` to nil to suppress usage of saved
  arguments unless explicitly specified.