fix: Suppress --override-binary when not supported by difft
The `--override-binary` is a new feature introduced in difftastic-0.65.0.  Use
the difftastic executable version to determine whether to display the infix in
transient as well as remove it from transient values when not supported.  The
latter is for a case when the transient value is saved on a system that has
difftastic 0.65.0 installed but later run with a `difftastic-executable` that
is in a lower version (e.g., on a different system).